Fuse's attr.Size is a uint64. Cast to that type.
This test is failing because testify's assert.Equal wants the types of
both to be the same.  So the uint32 LHS will never equal the uint64 RHS,
even if they have the same value.

The other option would be to remove the cast and use assert.EqualValues
but that doesn't seem as clean: We know what type the LHS needs to be,
so doing the cast explicitly seems best.

+++ fs_test.go
@@ -102,7 +102,7 @@ func (suite *FsTestSuite) TestFileAttrs() {
 		attr, status := suite.fs.GetAttr(c.filename, fuseContext)
 		assert.Equal(fuse.OK, status, "Expected %v attr status to be fuse.OK", c.filename)
 		assert.Equal(c.mode, attr.Mode, "Expected %v mode %#o, was %#o", c.filename, c.mode, attr.Mode)
-		assert.Equal(uint32(len(c.content)), attr.Size, "Expected %v size to match", c.filename)
+		assert.Equal(uint64(len(c.content)), attr.Size, "Expected %v size to match", c.filename)
 	}
 }
